Analyze after save
Run analysis after a save event.

Refactored context to dispatch on params.
Params a generally distinct enough to
dispatch json requests on for this protocol.

Also, this makes it easier to write utility
classes, since we don't have to account
for the entire Union of all possible requests
in every class.